refactor to only inline notification
